erts: Merge cons and boxed tags into one primary tag
Freeing up one whole primary tag by moving cons pointers into the boxed
pointer. We diffirentiate inbetween cons and boxed by checking what if
the boxed pointer points to a header word or not. If it points to a header
(i.e. something starting with 00) we know that it is a boxed term, if it
points to anything else (i.e. 01,11,10) we know that it is a cons cell.