erts: Make the gc compact {} to an immidiate
When doing a gc all 0-arity tuples are compacted to an immidiate. This
is useful because:
  1. it saves some memory.
  2. it guarantees that all boxed terms are atleast 2 words during a gc

Number 2 above is needed when we want to combine cons and boxed primary
tags into the same tag and therefore use the cons strategy for installing
forwarding pointers in the heap.

Ideally we want to always use the immidiate for 0-arity tuples, but this
turned out to be quite complex to achieve so for now we only compact at gc.