Copy old terms directly to old heap at fullsweep
This speeds up rapid heap growth by up to 40%. We still have to
fix so that the old heap is adjusted after the fullsweep. Also
if allocation of the old heap fails, we might want to just copy
all to new heap. Also gc of literals assumes that a fullsweep
does not leave a old heap, this has to be fixed.