erts: Copy old terms to old heap at fullsweep
This commit makes it so that terms that survive the fullsweep are
directly copied to the old heap rather than being copied to the new
heap. This will delay the next minor gc and since the data on the old
is already there is does not have to be copied again.

In importain detail is that the growing/shrinking of the new heap is
exactly as it was before. i.e the terms on the old heap are counted
as if they were on the new heap when figuring out if we want to shrink
or grow the new heap.

NOTE: This is a rewrite that is not tested and not complete, still have to
do the grow/schrink stuff and run the emulator.