erts: Prefetch Arg(0) to a register in beam_emu
This change is inspiered by what the LuaJIT vm does in
it's x86 interpreter. The idea is that the OOO of the x86
should do the load while the dispatch is happening so that
it does not have to be done in the actual instruction.

Basic idea:

Change
    OpCase(test_f): {
      Eterm arg = Arg(0);
      ... do things ...
      I += 2;
      Goto(*I);
    }

To
    OpCase(test_f): {
      Eterm arg = arg0;
      ... do things ...
      I += 2;
      arg0 = Arg(0);
      Goto(*I);
    }

As long as arg0 is placed in a register, this should
be an optimization.

One important change that is left to be made is that
we want values that are always used to be in arg0 so
that the load is always needed. Right now very often
the fail label ends up in arg0 which kind of nullifies
the optimization.