Implement relative f jumps
After loading all OTP modules, the largest relative offset
was 62802, which means that we cannot use 16 bits for packing
the f label, instead we have to use 32 bits. With 32 bits
modules can be 32 GBs large before the relative jump becomes
too large.

An idea I had with this is that if we inline the fail address
into the address word, we could jump make all instructions with
a fail label, emit a jump instruction that is skipped if the
error case is not taken. i.e.

Before:
    is_integer_fx f(FFFFFFFFFFFFFFFA) x(0)

After
    is_integer_x x(0)
    jump_f f(FFFFFFFFFFFFFFFA)

This would result in code that is as large, however we would be able
to place x(0) in the Arg(0) prefetch and the the fail label in there
as well. For some instructions the required size could even decrease.

The downside I guess is that is_integer would become three branches
instead of two when the fail label is taken.

This idea is also inspired by how the LuaJIT does it's fail label
dispatching.