Test to use lz4 compression for ets
Result:
  - lz4 is about 2x slower than doing t2b
  - decompression is about the same
  - lz4 compression factor is about 60%
  - t2b is about 20%

So, this is not worth it. The implementation does a copy_struct
and then does the lz4 compression. This since lz4 does not have
a nice streaming API for non-contigious memory chunks. Basically
the encode procedure becomes 3 pass instead of being 2 pass and
it is not worth it.

Some quick measurements show that we can reduce the size of ets
data by another 20% if we compress the t2b bytes.