erts: Fix nl mapping in to_erl terminal
ttysl assumes that a nl is only a nl and not a cr+nl so
we should set the to_erl tty in the same mode otherwise
they will be confused and out of sync.

I'm assuming that there was a reason why it was enabled
to start with but I can't think of a reason that would
still be applicable.