dialyzer: Assign better start locations to messages
For example, instead of pinpointing ":" in "m:f(", the leftmost
location of "m:f" is used in messages. Another example: instead of
pinpointing "=" in a matching, the start of the left operand is used.

This change assumes that erl_expand_records updates the locations of
record field initializations with the location of the record creation
or update.