Implement primary metadata
This commit moves the merge of primary metadata earlier
so that a better merge order is used. It also moves
the lookup of the primary config to be a bit earlier in
order to get access to the config without doing multiple
ets lookups.

The commit also renames it to primary metadata.