dialyzer: Use persistent_term for code
As an alternative to ETS, persistent_term can be used for storing Core
Erlang code.

An option '--use_persistent_term' has been added. The default is to
use ETS.

The memory consumption is often increased by as much as 25 to 30 per
cent when using persistent_term.

The execution time is often decreased by 5 to 10 per cent when using
persistent_term.