docgen: Make docs collapse better on small screens
This commit makes padding and margins smaller when the screen size
becomed smaller and when it is small enough it moves the navigation
to a collapseable topbar instead.

To achieve this the top-level layout has been changed from using
and fixed placed div for the navbar to use css grid instead.

When on a small screen the index pages' navigation is permanently
expanded to that the user does not have to expand it.