erts: Fix trace running memory leak
If a process is being tracing using 'running' trace and
at the same time polled using erlang:process_info, then
the memory allocated for the tracing will never trigger
a GC as there are no GC tests for when only signals/sys
tasks are run.

This commit moves the GC check so that it is done earlier
in order to catch any memory allocated when executing
signals/sys tasks.

The bug is triggered by running etop towards an idle node.