logger: Print to stderr if printing to stdout failed
If we cannot print to standard_io we try to instead print
to standard_error just to make sure that the error log message
is printed.

We know that group_leader will always point to the kernel application
master and thus we will be forwarded to 'user' eventutally, so we
can short circuit there. As a bonus we also now get to know if
user or user_drv has crashed and the io:put_chars call can fail.