compile: fix function/type/callback status
this commit fixes ill-detected functions/callbacks/types that were
`hidden` but set to `none`. the issue was the processing of the
`#docs.doc` and `#docs.doc_status`, where `#docs.doc` can be set to
`none` and, in this case, one cannot assume that `#docs.doc_status` is
also `none`, since it can be `hidden`. essentially, one needs to look at
the `#docs.doc_status` before checking whether there is some
documentation set in `#docs.doc`